Compute when graph is dirty, or the node last evaluated with is different.
There are cases where a Graph can be kClean, but evaluation needs to be
triggered again (hopefully looking into the cache for values).
For example, construct a graph with nodes A -> B -> C, then evaluate the
graph at C, next evaluate the graph at B. The Graph will still be clean,
but we wish to get the result of the graph at a different node (B for
example).