Refactor how geometry (buffers) are created and filled.
This update allows "deformers" (AKA lens distortion) to deform the
vertex positions using internal OCG parameters.

We can detect if the topology of the mesh is dirty or if the
vertices are dirty, or both, and then re-build the geometry as
needed (taking care of deformers at the same time).

Renamed
"image_plane_geometry_buffer.cpp/h" to "image_plane_geometry.cpp/h"
because the source creates geometry, not just buffers for the GPU/OpenGL.

Updating to latest OCG commit.