Add "uniqueNodeHash" attribute on all comp nodes at creation time.
The intention of this attribute is to be used to uniquely identify
the node in a large list of nodes created by anybody at any time.

Further, the idea is to use these unique hashes in a dynamic file path
to ensure any 2 nodes do not conflict. We will use this to create a
disk cached image sequence for each unique node. If this file system
is shared on a network, or a single user with multiple Maya sessions,
file path collisions are (practically) guaranteed to never happen.