Write node does not have file name attribute.
This is to stop the browse button from being used in the
Attribute Editor. Ideally, this attribute would still be marked
as 'used as filename', but the Attribute Editor buttons would be
re-created.

The 'used as filename' is used when "archiving" a Maya scene and can be
very helpful, however perhaps we don't actually want or need an
output node for this.