Add Disk Cache OCG node system to Maya ocgImageRead node.
The disk cache feature is an alternate "proxy" plate file name that
can be used. This file path that will be read and used rather than
the original values on the read node, if "diskCacheEnable" is enabled.

The alternate plate can be used to save memory by reducing the image
bit depth of the number of pixels to be written out.